requirements.txt
streamlit==1.39.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
pymongo==4.10.1
qdrant-client==1.12.1
langchain==0.3.2
langchain-qdrant==0.2.0
langchain-google-genai==1.0.5
langchain-community==0.3.2
google-generativeai==0.8.3
PyPDF2==3.0.1
python-dotenv==1.0.1
tqdm==4.66.5
